fix some warnings

+++ util/testing_harness.cpp
@@ -102,7 +102,7 @@ int main(int argc, char *argv[]) {
 	char *file_name = argv[1];
 
 	parse_json(file_name, parse_frame);
-	printf("Total frame:  %d, Pipeline frame rate: %d, video frame rate: %d\n", frame_count, FRAME_RATE, frame_rate);
+	printf("Total frame:  %ld, Pipeline frame rate: %d, video frame rate: %d\n", frame_count, FRAME_RATE, frame_rate);
 	
 	/* check if the frame rate is valid, to be valid the frame_rate has to be smaller or equal to FRAME_RATE and to be a power of 2 as well */
 	if (FRAME_RATE > frame_rate || (frame_rate % 2)) {
@@ -148,7 +148,7 @@ int main(int argc, char *argv[]) {
 	free(cur_frame);
 	free(buf_frame);
 	free(background);
-	for (int i = 0; i < frame_count; i++) {
+	for (unsigned long i = 0; i < frame_count; i++) {
 		free(frames_ptr[i]);
 	}
 	free(frames_ptr);
@@ -222,7 +222,7 @@ static void parse_frame(json_value* value, int depth) {
 	}
 
 	json_value *frames;
-	for (int i = 0; i < value->u.object.length; i++) {
+	for (unsigned int i = 0; i < value->u.object.length; i++) {
 		if (!strcmp(value->u.object.values[i].name, "frames"))
 			frames = value->u.object.values[i].value;
 	}
@@ -234,7 +234,7 @@ static void parse_frame(json_value* value, int depth) {
 	frame_count = frames->u.array.length;
 	frames_ptr = (double **)malloc( frame_count * sizeof(double *));	
 
-	for (int i = 0; i < frame_count; i++) {
+	for (unsigned long i = 0; i < frame_count; i++) {
 
 		double* frame_ptr = (double *)malloc(RESOLUTION * sizeof(double));
 		frames_ptr[i] = frame_ptr;
@@ -244,7 +244,7 @@ static void parse_frame(json_value* value, int depth) {
 			fprintf(stderr, "Invalid frame\n");
 			exit(1);
 		}
-		for (int j = 0; j < frame->u.array.length; j++) {
+		for (unsigned int j = 0; j < frame->u.array.length; j++) {
 			json_value* value_dbl = frame->u.array.values[j];
 			if (value_dbl->type != json_double) {
 				fprintf(stderr, "Invalid data format\n");
@@ -328,10 +328,10 @@ static void get_background(uint8_t* frame, unsigned long frame_count) {
  * @param depth a counter that counts how deep the recursion goes
  */
 static void read_config(json_value* value, int depth) {	
-	for (int i = 0; !depth && i < value->u.object.length; i++)  {
+	for (unsigned int i = 0; !depth && i < value->u.object.length; i++)  {
 		read_config(value->u.object.values[i].value, depth + 1);
 	}
-	for (int i = 0; depth == 1 && i < value->u.object.length; i++) {
+	for (unsigned int i = 0; depth == 1 && i < value->u.object.length; i++) {
 		json_value* temp = value->u.object.values[i].value;
 		int temp_value = temp->u.integer;	
 		switch (str2int(value->u.object.values[i].name)) {
